Fix issue with long filepath on Windows
