+++ test/Main.hs
@@ -13,7 +13,6 @@ import Algebra.Graph.Test.NonEmpty.Graph
 import Algebra.Graph.Test.Relation
 import Algebra.Graph.Test.Relation.SymmetricRelation
 import Algebra.Graph.Test.Bipartite.Undirected.AdjacencyMap
-import Algebra.Graph.Test.Bipartite.Undirected.AdjacencyMap.Algorithm
 import Algebra.Graph.Test.Undirected
 import Data.Graph.Test.Typed
